third_party/WebKit/LayoutTests/platform/android/bluetooth/requestDevice/chooser/device-removed-expected.txt

This is a testharness.js-based test.
FAIL Devices are added and removed. Chooser should not get a signal that the device was removed and promise should be rejected when a removed device is selected. assert_equals: Unexpected Error Name: expected "NotFoundError" but got "Error"
Harness: the test ran to completion.
